updated accountdelegation query & graphql-sdk types
