+++ packages/explorer/src/components/RebondForm/index.js
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { Form } from 'react-final-form'
-import Confetti from 'react-dom-confetti'
+import Confetti from '../Confetti'
 import { withProp } from '../../enhancers'
 import Button from '../Button'
 import InlineAccount from '../InlineAccount'
@@ -26,23 +26,10 @@ const RebondForm: React.StatelessFunctionalComponent<RebondFormProps> = ({
   values,
   ...props
 }) => {
-  const confetti = (
-    <Confetti
-      active={submitSucceeded}
-      config={{
-        angle: 90,
-        spread: 197,
-        startVelocity: 45,
-        elementCount: 50,
-        decay: 0.9,
-      }}
-    />
-  )
-
   if (submitSucceeded) {
     return (
       <React.Fragment>
-        {confetti}
+        <Confetti active={submitSucceeded} />
         <p>
           Sucess! {`${amount} LPT are now bonded back to ${delegateAddress}`}
         </p>
