+++ packages/explorer/src/views/Transcoders/enhance.js
@@ -5,6 +5,7 @@ import gql from 'graphql-tag'
 import {
   connectCurrentRoundQuery,
   connectProtocolQuery,
+  connectTranscodersQuery,
   connectToasts,
 } from '../../enhancers'
 import { MathBN } from '../../utils'
@@ -76,61 +77,6 @@ const connectMeDelegatorTranscoderQuery = graphql(MeDelegatorTranscoderQuery, {
   options: ({ match }) => ({
     pollInterval: 5 * 1000,
     variables: {},
-  }),
-})
-
-const TranscodersQuery = gql`
-  fragment TranscoderFragment on Transcoder {
-    id
-    active
-    ensName
-    status
-    lastRewardRound
-    rewardCut
-    feeShare
-    pricePerSegment
-    pendingRewardCut
-    pendingFeeShare
-    pendingPricePerSegment
-    totalStake
-    rewards {
-      rewardTokens
-      round {
-        id
-      }
-    }
-  }
-
-  query TranscodersQuery {
-    transcoders {
-      ...TranscoderFragment
-    }
-  }
-`
-
-const connectTranscodersQuery = graphql(TranscodersQuery, {
-  props: ({ data, ownProps }) => {
-    let { transcoders, ...queryData } = data
-    // Filter by registered transcoders
-    // TODO: Use graphql variables instead when The Graph supports them
-    if (transcoders) {
-      transcoders = transcoders.filter(t => t.status === 'Registered')
-    }
-    return {
-      ...ownProps,
-      transcoders: {
-        ...queryData,
-        data: transcoders || [],
-      },
-    }
-  },
-  options: ({ match }) => ({
-    pollInterval: 10 * 1000,
-    variables: {
-      skip: 0,
-      first: 100,
-    },
-    fetchPolicy: 'cache-and-network',
   }),
 })
